+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Changes in v1.11.6
 
+- Added a `remove all` button to `trading card manager`
+  - Allows users to remove all their active trading card listings on the Steam market
+- Better icons for `trading card manager` and its buttons
 - Fixed an issue where the trading card manager list wasn't updated if no cards were available
 
 <!-- 1.11.5 -->